Version 0.2.0
This commit has breaking changes. The interface moves from `Alphabet` to
`Phonetic` and an `Alphabet` is now represented by and `Enum`.
This `Enum` is auto-generated based on the files that are available in
`alphabets/`. This way, we leverage Rust's typesystem better while at
the same time we're able to document the `Enum` and thus the available
alphabets in an automated way. Since rust enums don't allow hypens, all
hyphens are now underscores in alphabet names.